Added very basic visual components
Board is now has a Sprite, once the game runs, it creates 40 Waypoints that represent the different tile locations (if u want to see the position of these, go to the Waypoint Prefab and turn on the SpriteRenderer).

The player can move from tile to tile, do not ask why it is Hank atm.