Improve player class & organise main scene hierarchy
I have reworked the player class to allow for backwards moves, as there are certain action cards that make you move backwards. But there are also action cards that force you to go clockwise even if the space is behind you so there is a toggle to only go forward, this took much longer than I thought it would 😅. I have also added documentation for all the key methods.