add rolling a double logic, add property for sale UI placeholder, add purchase and auction methods, rework player class to allow for bot overriding decision points, remove human class
Bots can play against themselves now. I realised the human class is not needed, the bot class overrides the methods where the player has to click a button and that is where it will manually decides which option is best. UI is just a placeholder atm and buy and auction methods are not implemented.